synq: split parser_macros.c into three focused files
parser_macros.c had three unrelated concerns in one file:
- Macro expansion pipeline (arg scanning, callbacks, template expansion)
- Span resolution and traceback (layer navigation)
- Per-node extent tracking hooks (AST builder machinery)

Split into:
- parser_macros.c (726 lines) — macro expansion pipeline
- parser_spans.c (250 lines) — span resolution + traceback
- parser_extents.c (108 lines) — extent hooks (operates on SynqParseCtx*)
