+++ syntaqlite-syntax/src/embedded_sources.rs
@@ -49,6 +49,8 @@ pub const RUNTIME_CSRC: &[(&str, &str)] = &[
     ),
     ("parser.c", include_str!("../csrc/parser.c")),
     ("parser_macros.c", include_str!("../csrc/parser_macros.c")),
+    ("parser_spans.c", include_str!("../csrc/parser_spans.c")),
+    ("parser_extents.c", include_str!("../csrc/parser_extents.c")),
     ("parser_dump.c", include_str!("../csrc/parser_dump.c")),
     (
         "parser_internal.h",
